Fix address update of already moved memory
When an object moved to a 'to' space, collector set a forward pointer inside
the old object to a new location. Then, if mark happens on an already moved
object, we just update incoming pointer with the forward pointer. See heap_mark.
Also, memory for a frame-local data is allocated from a heap - this
changes need to implement closures (will be display-based).