Optimize closure allocation
By defining table with safe builtin functions, I've made some simple
optimizations which allows to avoid closure allocations. It works when
we know, that env will not be overwriten so we don't have to allocate
closure for continuation. I hope to extend it not only for builtins but
for lambda too. Anyway, code generator will be rewriten in future in
order to use more complex optimisations.